Move meta feed helpers into own file
