+++ index.js
@@ -29,6 +29,7 @@ const { fromMessageSigil, isBendyButtV1FeedSSBURI } = require('ssb-uri2')
 const buildGroupId = require('./lib/build-group-id')
 const AddGroupTangle = require('./lib/add-group-tangle')
 const publishAndPrune = require('./lib/prune-publish')
+const MetaFeedHelpers = require('./lib/meta-feed-helpers')
 
 module.exports = {
   name: 'tribes2',
@@ -42,218 +43,12 @@ module.exports = {
   // eslint-disable-next-line no-unused-vars
   init(ssb, config) {
     const addGroupTangle = AddGroupTangle(ssb)
-
-    function findOrCreateAdditionsFeed(cb) {
-      const details = {
-        purpose: 'group/additions',
-        feedFormat: 'classic',
-      }
-
-      ssb.metafeeds.findOrCreate(details, cb)
-    }
-
-    function findEmptyGroupFeed(rootId, cb) {
-      let found = false
-      pull(
-        ssb.metafeeds.branchStream({ root: rootId, old: true, live: false }),
-        pull.filter((branch) => branch.length === 4),
-        pull.map((branch) => branch[3]),
-        // only grab feeds that look like group feeds
-        pull.filter((feed) => feed.recps && feed.purpose.length === 44),
-        paraMap((feed, cb) => {
-          ssb.db.query(
-            where(author(feed.id)),
-            count(),
-            toCallback((err, count) => {
-              // prettier-ignore
-              if (err) return cb(clarify(err, 'Failed to count messages in group feed'))
-
-              if (count === 0) {
-                // we're only interested in empty feeds
-                return cb(null, feed)
-              } else {
-                return cb()
-              }
-            })
-          )
-        }),
-        pull.filter(),
-        pull.unique('id'),
-        pull.take(1),
-        pull.drain(
-          (emptyFeed) => {
-            found = true
-            return cb(null, emptyFeed)
-          },
-          (err) => {
-            if (err) cb(clarify(err, 'Failed to find empty group feed'))
-            if (!found) cb()
-          }
-        )
-      )
-    }
-
-    function findOrCreateFromSecret(secret, rootId, cb) {
-      const recps = [
-        { key: secret.toBuffer(), scheme: keySchemes.private_group },
-        // encrypt to myself to be able to get back to the group without finding the group/add-member for me (maybe i crashed before adding myself)
-        rootId,
-      ]
-
-      const groupFeedDetails = {
-        purpose: secret.toString(),
-        feedFormat: 'classic',
-        recps,
-        encryptionFormat: 'box2',
-      }
-
-      ssb.metafeeds.findOrCreate(groupFeedDetails, (err, groupFeed) => {
-        if (err) return cb(clarify(err, 'Failed to find or create group feed'))
-        cb(null, groupFeed)
-      })
-    }
-
-    function findOrCreateGroupFeed(input = null, cb) {
-      const inputSecret = Buffer.isBuffer(input) ? new SecretKey(input) : input
-
-      ssb.metafeeds.findOrCreate(function gotRoot(err, root) {
-        // prettier-ignore
-        if (err) return cb(clarify(err, 'Failed to find or create root feed when finding or creating group feed'))
-
-        // 1. publish. we know the secret and should just findOrCreate the feed
-        // 2. create. we don't have a secret yet but can make or find one
-        //   2.1. try to find an empty feed in case we've crashed before. use the secret from that one
-        //   2.2. if we can't find an empty feed, make a new secret and findOrCreate
-
-        if (inputSecret) {
-          return findOrCreateFromSecret(inputSecret, root.id, cb)
-        } else {
-          findEmptyGroupFeed(root.id, (err, emptyFeed) => {
-            // prettier-ignore
-            if (err) return cb(clarify(err, 'Failed to find empty group feed when finding or creating a group'))
-
-            if (emptyFeed) {
-              return cb(null, emptyFeed)
-            } else {
-              const newSecret = new SecretKey()
-              findOrCreateFromSecret(newSecret, root.id, cb)
-            }
-          })
-        }
-      })
-    }
-
-    function secretKeyFromString(string) {
-      return new SecretKey(Buffer.from(string, 'base64'))
-    }
-
-    function createGroupWithoutMembers(myRoot, cb) {
-      findOrCreateGroupFeed(null, function gotGroupFeed(err, groupFeed) {
-        // prettier-ignore
-        if (err) return cb(clarify(err, 'Failed to find or create group feed when creating a group'))
-
-        const secret = secretKeyFromString(groupFeed.purpose)
-
-        const recps = [
-          { key: secret.toBuffer(), scheme: keySchemes.private_group },
-          myRoot.id,
-        ]
-
-        const content = {
-          type: 'group/init',
-          groupKey: secret.toString('base64'),
-          tangles: {
-            group: { root: null, previous: null },
-          },
-        }
-        if (!initSpec(content)) return cb(new Error(initSpec.errorsString))
-
-        ssb.db.create(
-          {
-            keys: groupFeed.keys,
-            content,
-            recps,
-            encryptionFormat: 'box2',
-          },
-          (err, groupInitMsg) => {
-            // prettier-ignore
-            if (err) return cb(clarify(err, "couldn't create group root message"))
-            return cb(null, { groupInitMsg, groupFeed, myRoot })
-          }
-        )
-      })
-    }
-
-    /** more specifically: a group that has never had any members. i.e. either
-     * 1. newly created but tribes2 crashed before we had time to add ourselves to it. in that case find and return it. if we can't find such a group then
-     * 2. freshly create a new group, and return */
-    function findOrCreateGroupWithoutMembers(cb) {
-      ssb.metafeeds.findOrCreate(function gotRoot(err, myRoot) {
-        // prettier-ignore
-        if (err) return cb(clarify(err, 'Failed to find or create root feed when creating a group'))
-
-        // find groups without any group/add-member messages
-        let foundMemberlessGroup = false
-        pull(
-          ssb.db.query(
-            where(and(isDecrypted('box2'), type('group/init'))),
-            toPullStream()
-          ),
-          pull.asyncMap((rootMsg, cb) => {
-            let foundMember = false
-            pull(
-              ssb.db.query(
-                where(and(isDecrypted('box2'), type('group/add-member'))),
-                toPullStream()
-              ),
-              pull.filter(
-                (msg) =>
-                  msg?.value?.content?.root === fromMessageSigil(rootMsg.key)
-              ),
-              pull.take(1),
-              pull.drain(
-                () => {
-                  foundMember = true
-                  return cb(null, false)
-                },
-                (err) => {
-                  if (err) return cb(err)
-                  else if (!foundMember) cb(null, rootMsg)
-                }
-              )
-            )
-          }),
-          pull.filter(Boolean),
-          pull.take(1),
-          pull.drain(
-            (rootMsg) => {
-              foundMemberlessGroup = true
-              ssb.metafeeds.advanced.findById(
-                rootMsg.value.author,
-                (err, groupFeed) => {
-                  // prettier-ignore
-                  if (err) return cb(clarify(err, "failed finding details of the memberless group feed"))
-
-                  rootMsg.value.content = rootMsg.meta.originalContent
-                  return cb(null, {
-                    groupInitMsg: rootMsg,
-                    groupFeed,
-                    myRoot,
-                  })
-                }
-              )
-            },
-            (err) => {
-              // prettier-ignore
-              if (err) return cb(clarify(err, "errored trying to find potential memberless feed"))
-              else if (!foundMemberlessGroup) {
-                return createGroupWithoutMembers(myRoot, cb)
-              }
-            }
-          )
-        )
-      })
-    }
+    const {
+      secretKeyFromString,
+      findOrCreateAdditionsFeed,
+      findOrCreateGroupFeed,
+      findOrCreateGroupWithoutMembers,
+    } = MetaFeedHelpers(ssb)
 
     function create(opts = {}, cb) {
       if (cb === undefined) return promisify(create)(opts)
